moved the class to correct package and deleted obsolete class
